gateway: mechanical split — move the seal-reasoning handler out of native_bridge
native_bridge.py reached 1030 physical lines with #772, tripping the
999-line repo-layout gate on main and every branch cut from it. Purely
mechanical relief: the seal_reasoning_content handler body moves
verbatim to native_reasoning.py (which already owns the bridge-side
reasoning-history handling) as seal_reasoning_carrier_content, taking
the attempt accounting as its one dependency; the bridge method becomes
a one-line delegator. Zero behavior change — no logic, message, or
error-shape edits. #772's lane owns this file's recent semantics; the
split deliberately avoids touching them.

native_bridge.py: 1030 -> 977 lines.
